--HG--
rename : js/src/js.cpp => js/src/js.c
rename : js/src/jsapi.cpp => js/src/jsapi.c
rename : js/src/jsarena.cpp => js/src/jsarena.c
rename : js/src/jsarray.cpp => js/src/jsarray.c
rename : js/src/jsatom.cpp => js/src/jsatom.c
rename : js/src/jsbool.cpp => js/src/jsbool.c
rename : js/src/jscntxt.cpp => js/src/jscntxt.c
rename : js/src/jscpucfg.cpp => js/src/jscpucfg.c
rename : js/src/jsdate.cpp => js/src/jsdate.c
rename : js/src/jsdbgapi.cpp => js/src/jsdbgapi.c
rename : js/src/jsdhash.cpp => js/src/jsdhash.c
rename : js/src/jsdtoa.cpp => js/src/jsdtoa.c
rename : js/src/jsemit.cpp => js/src/jsemit.c
rename : js/src/jsexn.cpp => js/src/jsexn.c
rename : js/src/jsfile.cpp => js/src/jsfile.c
rename : js/src/jsfun.cpp => js/src/jsfun.c
rename : js/src/jsgc.cpp => js/src/jsgc.c
rename : js/src/jshash.cpp => js/src/jshash.c
rename : js/src/jsinterp.cpp => js/src/jsinterp.c
rename : js/src/jsinvoke.cpp => js/src/jsinvoke.c
rename : js/src/jsiter.cpp => js/src/jsiter.c
rename : js/src/jskwgen.cpp => js/src/jskwgen.c
rename : js/src/jslock.cpp => js/src/jslock.c
rename : js/src/jslog2.cpp => js/src/jslog2.c
rename : js/src/jslong.cpp => js/src/jslong.c
rename : js/src/jsmath.cpp => js/src/jsmath.c
rename : js/src/jsnum.cpp => js/src/jsnum.c
rename : js/src/jsobj.cpp => js/src/jsobj.c
rename : js/src/jsopcode.cpp => js/src/jsopcode.c
rename : js/src/jsparse.cpp => js/src/jsparse.c
rename : js/src/jsprf.cpp => js/src/jsprf.c
rename : js/src/jsregexp.cpp => js/src/jsregexp.c
rename : js/src/jsscan.cpp => js/src/jsscan.c
rename : js/src/jsscope.cpp => js/src/jsscope.c
rename : js/src/jsscript.cpp => js/src/jsscript.c
rename : js/src/jsstr.cpp => js/src/jsstr.c
rename : js/src/jsutil.cpp => js/src/jsutil.c
rename : js/src/jsxdrapi.cpp => js/src/jsxdrapi.c
rename : js/src/jsxml.cpp => js/src/jsxml.c
rename : js/src/prmjtime.cpp => js/src/prmjtime.c

+++ testing/tests/l10n/lib/Mozilla/CompareLocales.py
@@ -38,74 +38,13 @@
 'Mozilla l10n compare locales tool'
 
 import os
+import os.path
 import re
 import logging
 import Parser
 import Paths
 
-def __regify(tpl):
-  return tuple(map(re.compile, tpl))
-
-exceptions = [
-  # ignore langpack contributor section for mail and brownser
-  __regify(('mail|browser', 'defines.inc', 'MOZ_LANGPACK_CONTRIBUTORS')),
-  # ignore search engine order for browser
-  __regify(('browser', 'chrome\\/browser-region\\/region\\.properties',
-   'browser\\.search\\.order\.[1-9]')),
-  # ignore feed engine order for browser
-  __regify(('browser', 'chrome\\/browser-region\\/region\\.properties',
-   'browser\\.contentHandlers\\.types\.[0-5]'))]
-
-def __dont_ignore(tpl):
-  for mod, path, key in exceptions:
-    if mod.match(tpl[0]) and path.match(tpl[1]) and key.match(tpl[2]):
-      return False
-  return True
-
-fl_exceptions = [
-  # ignore search plugins
-  __regify(('browser', 'searchplugins\\/.+\\.xml')),
-  # ignore help images
-  __regify(('browser', 'chrome\\/help\\/images\\/[A-Za-z-_]+\\.png'))]
-
-def do_ignore_fl(tpl):
-  for mod, path in fl_exceptions:
-    if mod.match(tpl[0]) and path.match(tpl[1]):
-      return True
-  return False
-
 class FileCollector:
-  class Iter:
-    def __init__(self, path):
-      self.__base = path
-    def __iter__(self):
-      self.__w = os.walk(self.__base)
-      try:
-        self.__nextDir()
-      except StopIteration:
-        # empty dir, bad, but happens
-        self.__i = [].__iter__()
-      return self
-    def __nextDir(self):
-      self.__t = self.__w.next()
-      try:
-        self.__t[1].remove("CVS")
-      except ValueError:
-        pass
-      self.__t[1].sort()
-      self.__t[2].sort()
-      self.__i = self.__t[2].__iter__()
-    def next(self):
-      try:
-        leaf = self.__i.next()
-        path = self.__t[0] + '/' + leaf
-        key = path[len(self.__base) + 1:]
-        return (key, path)
-      except StopIteration:
-        self.__nextDir()
-        return self.next()
-      print "not expected"
-      raise StopIteration
   def __init__(self):
     pass
   def getFiles(self, mod, locale):
@@ -114,18 +53,77 @@ class FileCollector:
       fls[leaf] = path
     return fls
   def iterateFiles(self, mod, locale):
-    return FileCollector.Iter(Paths.get_base_path(mod, locale))
+    base = Paths.get_base_path(mod, locale)
+    cutoff  = len(base) + 1
+    for dirpath, dirnames, filenames in os.walk(base):
+      try:
+        # ignore CVS dirs
+        dirnames.remove('CVS')
+      except ValueError:
+        pass
+      dirnames.sort()
+      filenames.sort()
+      for f in filenames:
+        leaf = dirpath + '/' + f
+        yield (leaf[cutoff:], leaf)
 
-def collectFiles(aComparer):
-  'returns new files, files to compare, files to remove'
+def collectFiles(aComparer, apps = None, locales = None):
+  '''
+  returns new files, files to compare, files to remove
+  apps or locales need to be given, apps is a list, locales is a
+  hash mapping applications to languages.
+  If apps is given, it will look up all-locales for all apps for the
+  languages to test.
+  'toolkit' is added to the list of modules, too.
+  '''
+  if not apps and not locales:
+    raise RuntimeError, "collectFiles needs either apps or locales"
+  if apps and locales:
+    raise RuntimeError, "You don't want to give both apps or locales"
+  if locales:
+    apps = locales.keys()
+    # add toolkit, with all of the languages of all apps
+    all = set()
+    for locs in locales.values():
+      all.update(locs)
+    locales['toolkit'] = list(all)
+  else:
+    locales = Paths.allLocales(apps)
+  modules = Paths.Modules(apps)
   en = FileCollector()
   l10n = FileCollector()
-  for cat in Paths.locales.keys():
-    logging.debug(" testing " + cat+ " on " + str(Paths.modules))
-    aComparer.notifyLocales(cat, Paths.locales[cat])
-    for mod in Paths.modules[cat]:
+  # load filter functions for each app
+  fltrs = []
+  for app in apps:
+    filterpath = 'mozilla/%s/locales/filter.py' % app
+    if not os.path.exists(filterpath):
+      continue
+    l = {}
+    execfile(filterpath, {}, l)
+    if 'test' not in l or not callable(l['test']):
+      logging.debug('%s does not define function "test"' % filterpath)
+      continue
+    fltrs.append(l['test'])
+  
+  # define fltr function to be used, calling into the app specific ones
+  # if one of our apps wants us to know about a triple, make it so
+  def fltr(mod, lpath, entity = None):
+    for f in fltrs:
+      keep  = True
+      try:
+        keep = f(mod, lpath, entity)
+      except Exception, e:
+        logging.error(str(e))
+      if not keep:
+        return False
+    return True
+  
+  for cat in modules.keys():
+    logging.debug(" testing " + cat+ " on " + str(modules))
+    aComparer.notifyLocales(cat, locales[cat])
+    for mod in modules[cat]:
       en_fls = en.getFiles(mod, 'en-US')
-      for loc in Paths.locales[cat]:
+      for loc in locales[cat]:
         fls = dict(en_fls) # create copy for modification
         for l_fl, l_path in l10n.iterateFiles(mod, loc):
           if fls.has_key(l_fl):
@@ -133,11 +131,21 @@ def collectFiles(aComparer):
             aComparer.compareFile(mod, loc, l_fl)
             del fls[l_fl]
           else:
-            # file in locale, but not in en-US, remove?
-            aComparer.removeFile(mod, loc, l_fl)
+            if fltr(mod, l_fl):
+              # file in locale, but not in en-US, remove
+              aComparer.removeFile(mod, loc, l_fl)
+            else:
+              logging.debug(" ignoring %s from %s in %s" %
+                            (l_fl, loc, mod))
         # all locale files dealt with, remaining fls need to be added?
         for lf in fls.keys():
-          aComparer.addFile(mod,loc,lf)
+          if fltr(mod, lf):
+            aComparer.addFile(mod,loc,lf)
+          else:
+            logging.debug(" ignoring %s from %s in %s" %
+                          (lf, loc, mod))
+
+  return fltr
 
 class CompareCollector:
   'collects files to be compared, added, removed'
@@ -152,9 +160,6 @@ class CompareCollector:
       else:
         self.modules[loc] = [aModule]
   def addFile(self, aModule, aLocale, aLeaf):
-    if do_ignore_fl((aModule, aLeaf)):
-      logging.debug(" ignoring %s from %s in %s" % (aLeaf, aLocale, aModule))
-      return
     logging.debug(" add %s for %s in %s" % (aLeaf, aLocale, aModule))
     if not self.files.has_key(aLocale):
       self.files[aLocale] = {'missingFiles': [(aModule, aLeaf)],
@@ -169,9 +174,6 @@ class CompareCollector:
       self.cl[(aModule, aLeaf)].append(aLocale)
     pass
   def removeFile(self, aModule, aLocale, aLeaf):
-    if do_ignore_fl((aModule, aLeaf)):
-      logging.debug(" ignoring %s from %s in %s" % (aLeaf, aLocale, aModule))
-      return
     logging.debug(" remove %s from %s in %s" % (aLeaf, aLocale, aModule))
     if not self.files.has_key(aLocale):
       self.files[aLocale] = {'obsoleteFiles': [(aModule, aLeaf)],
@@ -180,10 +182,11 @@ class CompareCollector:
       self.files[aLocale]['obsoleteFiles'].append((aModule, aLeaf))
     pass
 
-def compare(testLocales=[]):
+def compare(apps=None, testLocales=None):
   result = {}
   c = CompareCollector()
-  collectFiles(c)
+  fltr = collectFiles(c, apps=apps, locales=testLocales)
+  
   key = re.compile('[kK]ey')
   for fl, locales in c.cl.iteritems():
     (mod,path) = fl
@@ -201,7 +204,7 @@ def compare(testLocales=[]):
       enTmp = dict(enMap)
       parser.read(Paths.get_path(mod, loc, path))
       for k,v in parser:
-        if not __dont_ignore((mod, path, k)):
+        if not fltr(mod, path, k):
           if enTmp.has_key(k):
             del enTmp[k]
           continue
@@ -219,7 +222,8 @@ def compare(testLocales=[]):
                          (k, Paths.get_path(mod, loc, path)))
           else:
             result[loc]['changed'] +=1
-      result[loc]['missing'].extend(filter(__dont_ignore, [(mod,path,k) for k in enTmp.keys()]))
+      result[loc]['missing'].extend(filter(lambda t: fltr(*t),
+                                           [(mod,path,k) for k in enTmp.keys()]))
   for loc,dics in c.files.iteritems():
     if not result.has_key(loc):
       result[loc] = dics
